refactor(cactus-web): extract shared RadioGroup/CheckBoxGroup functionality into Fieldset component
CACTUS-545
