+++ modules/cactus-web/src/RadioGroup/RadioGroup.tsx
@@ -1,17 +1,10 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import styled from 'styled-components'
 import { MarginProps, WidthProps } from 'styled-system'
 
-import { FieldProps, useAccessibleField } from '../AccessibleField/AccessibleField'
-import { TooltipAlignment } from '../AccessibleField/AccessibleField'
-import Box from '../Box/Box'
+import { FieldProps } from '../AccessibleField/AccessibleField'
 import Fieldset from '../Fieldset/Fieldset'
-import { cloneAll } from '../helpers/react'
-import Label from '../Label/Label'
 import RadioButtonField, { RadioButtonFieldProps } from '../RadioButtonField/RadioButtonField'
-import StatusMessage from '../StatusMessage/StatusMessage'
-import Tooltip from '../Tooltip/Tooltip'
 
 interface RadioGroupProps
   extends MarginProps,
@@ -22,9 +15,6 @@ interface RadioGroupProps
   defaultValue?: string | number
   required?: boolean
 }
-interface LabelWrapper {
-  alignTooltip?: TooltipAlignment
-}
 
 // This allows omitting the required `name` prop, since it'll be injected by the RadioGroup.
 type RadioGroupButtonProps = Omit<RadioButtonFieldProps, 'name' | 'required'>
@@ -33,49 +23,10 @@ const RadioGroupButton = React.forwardRef<HTMLInputElement, RadioGroupButtonProp
 )
 RadioGroupButton.displayName = 'RadioGroup.Button'
 
-type ForwardProps = {
-  name: string
-  disabled?: boolean
-  required?: boolean
-}
-
 const noop = () => undefined
 
 export const RadioGroup = React.forwardRef<HTMLFieldSetElement, RadioGroupProps>(
-  (
-    {
-      label,
-      children,
-      tooltip,
-      required,
-      defaultValue,
-      value,
-      onFocus,
-      onBlur,
-      autoTooltip = true,
-      disableTooltip,
-      alignTooltip = 'right',
-      ...props
-    },
-    ref
-  ) => {
-    const {
-      fieldId,
-      ariaDescribedBy,
-      labelId,
-      statusId,
-      name,
-      status,
-      statusMessage,
-      tooltipId,
-      disabled,
-    } = useAccessibleField({ ...props, tooltip })
-    const [showTooltip, setTooltipVisible] = React.useState<boolean>(false)
-
-    const forwardProps: ForwardProps = { name, required }
-    if (disabled === true) {
-      forwardProps.disabled = disabled
-    }
+  ({ children, defaultValue, value, ...props }, ref) => {
     const hasOnChange = !!props.onChange
     const cloneWithValue = (element: React.ReactElement, cloneProps: any) => {
       if (value !== undefined) {
@@ -98,79 +49,20 @@ export const RadioGroup = React.forwardRef<HTMLFieldSetElement, RadioGroupProps>
       }
       return React.cloneElement(element, cloneProps)
     }
-    children = cloneAll(children, forwardProps, cloneWithValue)
-
-    const handleFocus = React.useCallback(
-      (e: React.FocusEvent<HTMLFieldSetElement>) => {
-        onFocus?.(e)
-        if (!e.currentTarget.contains(e.relatedTarget as Node)) {
-          setTooltipVisible(() => true)
-        }
-      },
-      [onFocus, setTooltipVisible]
-    )
-    const handleBlur = React.useCallback(
-      (e: React.FocusEvent<HTMLFieldSetElement>) => {
-        onBlur?.(e)
-        if (!e.currentTarget.contains(e.relatedTarget as Node)) {
-          setTooltipVisible(() => false)
-        }
-      },
-      [onBlur, setTooltipVisible]
-    )
 
     return (
       <Fieldset
         {...props}
         ref={ref}
-        id={fieldId}
         role="radiogroup"
-        aria-describedby={ariaDescribedBy}
-        name={name}
-        disabled={disabled}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
+        forwardProps={['name', 'required']}
+        cloneWithValue={cloneWithValue}
       >
-        <Label id={labelId} as="legend">
-          <LabelWrapper alignTooltip={alignTooltip}>
-            {label}
-            {tooltip && (
-              <Tooltip
-                id={tooltipId}
-                label={tooltip}
-                disabled={disableTooltip ?? disabled}
-                forceVisible={autoTooltip ? showTooltip : false}
-              />
-            )}
-          </LabelWrapper>
-        </Label>
-        <Box mx={4} pt={3}>
-          {children}
-        </Box>
-        {status && (
-          <div>
-            <StatusMessage status={status} id={statusId}>
-              {statusMessage}
-            </StatusMessage>
-          </div>
-        )}
+        {children}
       </Fieldset>
     )
   }
 )
-
-const LabelWrapper = styled.span<LabelWrapper>`
-  flex-wrap: nowrap;
-  display: flex;
-  justify-content: ${(p) => (p.alignTooltip === 'right' ? 'space-between' : 'flex-start')};
-  padding-right: 8px;
-  ${Tooltip} {
-    position: relative;
-    right: 0;
-    bottom: 0;
-    padding-left: ${(p) => (p.alignTooltip === 'right' ? 0 : '8px')};
-  }
-`
 
 RadioGroup.displayName = 'RadioGroup'
 RadioGroup.propTypes = {
